last version 30/08/2024 05:34

+++ app/Http/Controllers/IncidentController.php
@@ -16,7 +16,7 @@ class IncidentController extends Controller
         return response()->json(['allIncidents'=>$incidents]);
 
     }
-    public function index($id)
+    public function index()
     {
         $user = JWTAuth::parseToken()->authenticate();
 
@@ -24,8 +24,7 @@ class IncidentController extends Controller
             return response()->json(['message' => 'Unauthenticated'], 401);
         }
 
-        $incidents =$user->incidents->where('building_id',$id)
-                    ->orWhere('component_id',$id)->get();
+        $incidents =$user->incidents()->get();
 
         event(new UserAction($user->id, 'viewed_incidents', 'User viewed incidents '));
 
@@ -35,8 +34,7 @@ class IncidentController extends Controller
     public function store(Request $request)
     {
 
-        Log::info("new incident data:");
-        Log::info($request->all());
+        //?worked
 
         $user = JWTAuth::parseToken()->authenticate();
 
@@ -57,8 +55,6 @@ class IncidentController extends Controller
             return response()->json(['errors' => $validator->errors()], 400);
         }
 
-        
-
         $incident = new Incident();
         $incident->title = $request->title;
         $incident->description = $request->description;
@@ -68,7 +64,6 @@ class IncidentController extends Controller
         $incident->component_id = $request->component_id;
         $incident->save();
 
-        // Dispatch event
         event(new UserAction($user->id, 'created_incident', 'User created an incident.'));
 
         return response()->json(['message'=>"New incident  created successfully ."]);
@@ -94,58 +89,21 @@ class IncidentController extends Controller
         return response()->json($incident);
     }
 
-    public function update(Request $request, $incidentId)
+    public function updateIncident(Request $request)
     {
+
+        // Log::info("incident update id");
+        // Log::info($request->id);
+        Log::info($request->all());
         $user = JWTAuth::parseToken()->authenticate();
 
         if (!$user) {
             return response()->json(['message' => 'Unauthenticated'], 401);
         }
 
-
-        $incident = Incident::find($incidentId);
-
-        if (!$incident) {
-            return response()->json(['message' => 'Incident not found'], 404);
-        }
-
-        $validator = Validator::make($request->all(), [
-            'title' => 'required|string|max:255|min:3',
-            'description' => 'required|string|max:255|min:3',
-            'status' => 'required|string|max:50',
-            'component_id' => 'required|exists:components,id',
-            //remember to remove the building_id
-            'building_id' => 'required|exists:buildings,id',
-            'user_id' => 'required|exists:users,id',
-        ]);
-
-        if ($validator->fails()) {
-            return response()->json(['errors' => $validator->errors()], 400);
-        }
-
-        $incident->title = $request->title;
-        $incident->description = $request->description;
-        $incident->status = $request->status;
-        $incident->component_id = $request->component_id;
-        $incident->building_id = $request->building_id;
-        $incident->user_id = $request->user_id;
-        $incident->save();
-
-        event(new UserAction($user->id, 'updated_incident', `User updated incident {$request->title} successfully.` ));
-
-        return response()->json(['message'=> `Incident {$request->title} updated successfully.`]);
-    }
-
-    public function destroyIncident(Request $request)
-    {
-        Log::info('incident delete id:');
-        Log::info($request->id);
-
-        $user = JWTAuth::parseToken()->authenticate();
-
-        if (!$user) {
-            return response()->json(['message' => 'Unauthenticated'], 401);
-        }
+        // $updatedIncident=$request->input('newIncident');
+        // Log::info('updatedIncident');
+        // Log::info($updatedIncident);
 
         $incident = Incident::find($request->id);
 
@@ -153,10 +111,59 @@ class IncidentController extends Controller
             return response()->json(['message' => 'Incident not found'], 404);
         }
 
+        $validator = Validator::make($request->all(), [
+            'id'=>'required',
+            'title' => 'required|string|max:255|min:3',
+            'description' => 'required|string|max:255|min:3',
+            'status' => 'required|string|max:50',
+            'component_id' => 'nullable|exists:components,id',
+            //remember to remove the building_id
+            'building_id' => 'nullable|exists:buildings,id',
+            'user_id' => 'required|exists:users,id',
+            'created_at' => 'required',
+            'updated_at' => 'required',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['errors' => $validator->errors()], 400);
+        }
+
+        $incident->id = $request->id;
+        $incident->title = $request->title;
+        $incident->description = $request->description;
+        $incident->status = $request->status;
+        $incident->user_id = $request->user_id;
+        $incident->component_id = $request->component_id;
+        $incident->building_id = $request->building_id;
+        $incident->created_at = $request->created_at;
+        $incident->updated_at = $request->updated_at;
+        $incident->save();
+
+        event(new UserAction($user->id, 'updated_incident', `User updated incident `.$request->title.`  successfully.` ));
+
+        return response()->json(['message'=> 'Incident '.$request->title.' updated  successfully. ']);
+    }
+
+    public function destroyIncident(Request $request)
+    {
+        $user = JWTAuth::parseToken()->authenticate();
+        
+        if (!$user) {
+            return response()->json(['message' => 'Unauthenticated'], 401);
+        }
+        // Log::info('IIIDDD');
+        
+        $incident = Incident::find($request->id);
+        
+        if (!$incident) {
+            Log::info('not found');
+            return response()->json(['message' => 'Incident not found'], 404);
+        }
+
         $incident->delete();
 
         event(new UserAction($user->id, 'deleted_incident', `User deleted incident {$incident->title} successfully.`));
 
-        return response()->json(['message' => `User deleted incident {$incident->title} successfully.`]);
+        return response()->json(['message' => 'Incident '.$incident->title.' deleted  successfully. ']);
     }
 }
